use long int for integers (oe).
git-svn-id: http://svn.delph-in.net/lkb/trunk@2117 3df82f5b-d43a-0410-af33-fce91db48ec5

+++ src/tsdb/c/tsdb.h
@@ -267,7 +267,7 @@ typedef struct tsdb_field {
 typedef struct tsdb_value {
   BYTE type;
   union {
-    int integer;
+    long int integer;
     char *identifier;
     char *string;
     char *date;
